requirements.txt
# CV Generator Requirements
# Core dependencies for generating PDF CVs from YAML templates

# Template engine for populating LaTeX templates
Jinja2>=3.1.0

# YAML parsing for loading CV data
PyYAML>=6.0

# Validation dependencies
jsonschema>=4.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# Optional: for enhanced LaTeX integration (if needed in future)
# Note: The actual PDF compilation is done by external LaTeX tools (pdflatex/latexmk)
# which need to be installed separately on the system